# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DoReMi - Plateforme scolaire numérique",
        "description": "Bienvenue dans la documentation de l'API **DoReMi**, une plateforme complète de gestion scolaire hybride.\\n\\n\n        ## Modules couverts :\\n\n        - Gestion des écoles et utilisateurs (élèves, enseignants, parents)\\n\n        - Cours en ligne et en présentiel\\n\n        - Suivi des notes et bulletins\\n\n        - Gestion des documents et devoirs\\n\n        - Actualités, emploi du temps, messagerie\\n\n        - Gestion des stages : demandes & offres\\n\\n\n        Cette documentation est à jour pour l'environnement de développement.",
        "termsOfService": "https://doremi-app.test/conditions",
        "contact": {
            "name": "Support Technique DoReMi",
            "email": "support@doremi-app.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/api/analytics/courses": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Obtenir les statistiques des cours",
                "operationId": "25bb1da482237738530326f7bf1e44a4",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (7d, 30d, 90d, 1y)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "30d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques des cours récupérées avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/analytics/users": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Obtenir les statistiques des utilisateurs",
                "operationId": "310c86edc71e7560a815600ca8f7deb0",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (7d, 30d, 90d, 1y)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "30d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques des utilisateurs récupérées avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/analytics/revenue": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Obtenir les statistiques de revenus",
                "operationId": "3b6aaa16d6dd6a617f5b36510b42f0bc",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (7d, 30d, 90d, 1y)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "30d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques de revenus récupérées avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/analytics/content": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Obtenir les statistiques du contenu",
                "operationId": "1dcb85eff5b365858e056b25043796b4",
                "responses": {
                    "200": {
                        "description": "Statistiques du contenu récupérées avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/generate": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Générer un rapport personnalisé",
                "operationId": "47ef8bede54dcaab9875aa3dc6e24215",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de rapport (courses, users, revenue, content)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (7d, 30d, 90d, 1y)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "30d"
                        }
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "description": "Format du rapport (json, csv)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rapport généré avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}